Fix race condition between monitoring backend and monitored backend when long SQL query is cancelled and SetLatch is not called because the monitored backend is interrupted before calling SetLatch().
In order to avoid such race condition, interrupts are held when the monitored
backend dump the progress state.

It could meanwhile happen that the query be cancelled even before reaching the
critical section: between signal received and call to HOLD_INTERRUPTS() in
HandleProgressRequest().

In order to avoid hanging the monitoring backend endlessly, a PROGRESS_TIMEOUT
has been added to waitLatch(). Default value is 5 seconds. If no response is
received during the PROGRESS_TIMEOUT, the monitoring backend will reply with
<backend timeout>